+++ lib/connection/response.py
@@ -16,6 +16,7 @@
 #
 #  Author: Mauro Soria
 
+import time
 import httpx
 
 from lib.core.settings import (
@@ -25,11 +26,12 @@ from lib.core.settings import (
     UNKNOWN,
 )
 from lib.parse.url import clean_path, parse_path
-from lib.utils.common import is_binary
+from lib.utils.common import get_readable_size, is_binary
 
 
 class BaseResponse:
     def __init__(self, response):
+        self.datetime = time.strftime("%Y-%m-%d %H:%M:%S")
         self.url = str(response.url)
         self.full_path = parse_path(self.url)
         self.path = clean_path(self.full_path)
@@ -53,6 +55,10 @@ class BaseResponse:
             return int(self.headers.get("content-length"))
         except TypeError:
             return len(self.body)
+
+    @property
+    def size(self):
+        return get_readable_size(self.length)
 
     def __hash__(self):
         return hash(self.body)